+++ CHANGES.rst
@@ -1,19 +1,30 @@
 ``sweetpotatopie`` Changelog
 ============================
 
+Unreleased
+----------
+
+- Add a field type:  colander.Money
+
+- Add remaiing widget types:  autocomplete, checkbox, checked, date,
+  checked_password, dateparts, datetime, hidden, input, input_csv, money,
+  paswword, textarea_csv, upload.
+
+
 0.2 (2013-04-11)
 ----------------
 
-- Add some added field types:  colander.Time, colander.Set
- 
+- Add some field types:  colander.Time, colander.Set
+
 - Add useful non-default widget types:  checkbox-choices, radio, rich text,
   select, textarea.
-  
+
 
 0.1.1 (2013-04-11)
 ------------------
 
 - Packaging fix (correct ReST rendering on PyPI).
+
 
 0.1 (2013-04-11)
 ----------------